Add a paragraph about attribute takeover
It mmight happen that already existing attributes shall be taken over by the FIG
to allow greater reach or provide better maintenance. This is in general
something that we allow and already thought about. Such attributes can be
published on one central page to make people more aware of them once the process
has been started of taking the attribute over.